+++ tests/e2e.ts
@@ -6,6 +6,7 @@ fixture("Svelte Persistent Storage")
 const cookieInput = Selector('#cookieInput'),
     localInput = Selector('#localInput'),
     sessionInput = Selector('#sessionInput'),
+    indexedInput = Selector('#indexedInput'),
     documentCookie = Selector('#documentCookie'),
     reloadButton = Selector('#reloadButton'),
     clearButton = Selector('#clearButton')
@@ -17,6 +18,7 @@ test("Initial state", async t => {
         .expect(cookieInput.value).eql("John")
         .expect(localInput.value).eql('Foo')
         .expect(sessionInput.value).eql('Bar')
+        .expect(indexedInput.value).eql('Hello')
         .expect(documentCookie.textContent).contains("sps-userName=%22John%22")
 })
 
@@ -54,4 +56,15 @@ test("Session storage", async t => {
         .click(clearButton)
         .click(reloadButton)
         .expect(sessionInput.value).eql("Bar")
+})
+
+test("IndexedDB storage", async t => {
+    await t
+        .expect(indexedInput.value).eql("Hello")
+        .typeText(indexedInput, " World")
+        .click(reloadButton)
+        .expect(indexedInput.value).eql("Hello World")
+        .click(clearButton)
+        .click(reloadButton)
+        .expect(indexedInput.value).eql("Hello")
 })