+++ tests/src/App.svelte
@@ -1,10 +1,11 @@
 <script>
-    import {persist, cookieStorage, localStorage, sessionStorage} from "../../src/index"
+    import {persist, cookieStorage, localStorage, sessionStorage, indexedDBStorage} from "../../src/index"
     import { writable } from "svelte/store"
 
     let cookieExample = persist(writable('John'), cookieStorage(), 'sps-userName')
     let localExample = persist(writable('Foo'), localStorage(), 'sps-action')
     let sessionExample = persist(writable('Bar'), sessionStorage(), 'sps-call')
+    let indexedDBExample = persist(writable('Hello'), indexedDBStorage(), 'sps-data')
 
     let cookie = ''
 
@@ -31,6 +32,11 @@
     <label>Enter a word: <input id="sessionInput" bind:value={$sessionExample}></label>
 </fieldset>
 
+<fieldset>
+    <legend>IndexedDB Storage</legend>
+    <label>Enter a word: <input id="indexedInput" bind:value={$indexedDBExample}></label>
+</fieldset>
+
 <button id="reloadButton" on:click={() => window.location.reload()}>Reload the page</button>
 
-<button id="clearButton" on:click={() => {cookieExample.delete(); localExample.delete(); sessionExample.delete()}}>Clear storages</button>
+<button id="clearButton" on:click={() => {cookieExample.delete(); localExample.delete(); sessionExample.delete(); indexedDBExample.delete()}}>Clear storages</button>